Thanks to python-ogre for the way to get around the tr1 errors. GCCXML can't handle the new C++ TR1 stuff yet

+++ wrappers/ogre/ogre.rake
@@ -1,6 +1,7 @@
 namespace :ogre do
 
-  OGRE_DOWNLOAD = "ogre-v1-6-2.tar.bz2"
+  OGRE_VERSION = "ogre-v1-6-2"
+  OGRE_DOWNLOAD = "#{OGRE_VERSION}.tar.bz2"
 
   desc "clean up ogre libs" 
   task :clean do
@@ -21,16 +22,20 @@ namespace :ogre do
   desc "setup libnoise for wrapping"
   task :setup => :bootstrap do
     cd File.join(OGRE_RB_ROOT , "tmp", "downloads") do
+      break if File.exists?(OGRE_DOWNLOAD)
       sh "wget http://downloads.sourceforge.net/ogre/#{OGRE_DOWNLOAD}"
       cp OGRE_DOWNLOAD, ".."
     end
 
     cd File.join(OGRE_RB_ROOT, "tmp") do
-      sh "tar jxvf #{OGRE_DOWNLOAD}"
+      system "tar jxvf #{OGRE_DOWNLOAD}"
     end
 
     cd File.join(OGRE_RB_ROOT, "tmp", "ogre") do
       install_to = File.join(OGRE_RB_ROOT, "lib", "ogre")
+      cmd = "patch -s -N -i  #{File.join(OGRE_RB_ROOT, "wrappers", "ogre", "patch", OGRE_VERSION + ".patch")} -p0"
+      puts "Patch command: #{cmd.inspect}"
+      sh cmd
       sh "./configure --prefix=#{install_to} && make && make install"
     end
   end
